examples/requirements.txt
requests>=2.28
websockets>=11.0
Pillow>=9.0
numpy>=1.23
